// model/createMatchTrial400Response.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateMatchTrial400Response {
    'code'?: CreateMatchTrial400Response.CodeEnum;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "CreateMatchTrial400Response.CodeEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateMatchTrial400Response.attributeTypeMap;
    }
}

export namespace CreateMatchTrial400Response {
    export enum CodeEnum {
        InvalidBodyFormat = <any> 'InvalidBodyFormat',
        InvalidParameters = <any> 'InvalidParameters'
    }
}