// model/getMatchEvaluation404Response.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class GetMatchEvaluation404Response {
    'code'?: GetMatchEvaluation404Response.CodeEnum;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "GetMatchEvaluation404Response.CodeEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return GetMatchEvaluation404Response.attributeTypeMap;
    }
}

export namespace GetMatchEvaluation404Response {
    export enum CodeEnum {
        MatchNotFound = <any> 'MatchNotFound',
        CompetitionNotFound = <any> 'CompetitionNotFound',
        TrialNotFound = <any> 'TrialNotFound',
        EvaluationNotFound = <any> 'EvaluationNotFound'
    }
}