// model/getMatchScore404Response.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class GetMatchScore404Response {
    'code'?: GetMatchScore404Response.CodeEnum;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "GetMatchScore404Response.CodeEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return GetMatchScore404Response.attributeTypeMap;
    }
}

export namespace GetMatchScore404Response {
    export enum CodeEnum {
        MatchNotFound = <any> 'MatchNotFound',
        CompetitionNotFound = <any> 'CompetitionNotFound',
        TrialNotFound = <any> 'TrialNotFound',
        ScoreNotFound = <any> 'ScoreNotFound'
    }
}