// model/getMatchTrial403Response.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class GetMatchTrial403Response {
    'code'?: GetMatchTrial403Response.CodeEnum;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "code",
            "baseName": "code",
            "type": "GetMatchTrial403Response.CodeEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return GetMatchTrial403Response.attributeTypeMap;
    }
}

export namespace GetMatchTrial403Response {
    export enum CodeEnum {
        CompetitionNotStarted = <any> 'CompetitionNotStarted',
        CompetitionNotInProgress = <any> 'CompetitionNotInProgress'
    }
}