// model/matchTrialEvaluation.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { RunnerStatus } from './runnerStatus';
import { ScalarOrVector } from './scalarOrVector';

/**
* Evaluation results
*/
export class MatchTrialEvaluation {
    'status': RunnerStatus;
    /**
    * Evaluation error information
    */
    'error'?: string;
    'objective'?: ScalarOrVector;
    'constraint'?: ScalarOrVector;
    /**
    * Auxiliary information for evaluation
    */
    'info'?: object;
    /**
    * Whether it is a feasible solution
    */
    'feasible'?: boolean;
    /**
    * Evaluation start date and time
    */
    'startedAt': Date;
    /**
    * Evaluation end date and time
    */
    'finishedAt': Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "RunnerStatus"
        },
        {
            "name": "error",
            "baseName": "error",
            "type": "string"
        },
        {
            "name": "objective",
            "baseName": "objective",
            "type": "ScalarOrVector"
        },
        {
            "name": "constraint",
            "baseName": "constraint",
            "type": "ScalarOrVector"
        },
        {
            "name": "info",
            "baseName": "info",
            "type": "object"
        },
        {
            "name": "feasible",
            "baseName": "feasible",
            "type": "boolean"
        },
        {
            "name": "startedAt",
            "baseName": "startedAt",
            "type": "Date"
        },
        {
            "name": "finishedAt",
            "baseName": "finishedAt",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return MatchTrialEvaluation.attributeTypeMap;
    }
}

export namespace MatchTrialEvaluation {
}