// model/matchTrialScore.ts
/**
 * OptHub REST API
 * OptHub Public REST API.
 *
 * The version of the OpenAPI document: 0.3.0
 * Contact: dev@opthub.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { RunnerStatus } from './runnerStatus';

/**
* Results of Score calculation
*/
export class MatchTrialScore {
    'status': RunnerStatus;
    /**
    * Score calculation error information
    */
    'error'?: string;
    /**
    * The value of the Score
    */
    'value'?: number;
    /**
    * Score calculation start date and time
    */
    'startedAt': Date;
    /**
    * Score calculation end date and time
    */
    'finishedAt': Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "RunnerStatus"
        },
        {
            "name": "error",
            "baseName": "error",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "number"
        },
        {
            "name": "startedAt",
            "baseName": "startedAt",
            "type": "Date"
        },
        {
            "name": "finishedAt",
            "baseName": "finishedAt",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return MatchTrialScore.attributeTypeMap;
    }
}

export namespace MatchTrialScore {
}